Cleanup raycast routines
Make this code actually readable again, though there's still room for improvement.

Also changed the wall optimisation method to draw each post on the same pass instead of on the next, which could cause problems with any added features in the future.

This optimisation is not very useful anymore, but it could be if the postsource pointer needed to point to something that would have to be calculated every frame without the "last" postsource, such as a table of wrapped vertical columns for a scrolling effect.