Temporary fix for random crashes on startup
Something is randomly causing a seg fault in CA_HuffExpand with the source pointer, and with more and more Tile8s, the more likely the crash is to occur.

As a temporary measure, until I find the source of the problem, NUMTILE8 is now 8 instead of 72 (since only 8 are being used for now).