Cleanup input manager & improve some functions
There's plenty of room for further improvement: IN_HandleEvent should take care of mouse events so that the button mask can be saved and used everywhere. This would easily allow mouse input to be grabbed by clicking in the window (and the mouse button cleared so it doesn't make you shoot).

ReadAnyControl can be merged into IN_ReadControl, and the latter can be setup to ignore mouse motion (could be useful for scrolling with the mouse wheel).

There's potential for setting up an array of keys that should be "ignored" on that frame, so that keys/buttons held down will be ignored until released. However, this would be tricky to get working because there's so much stuff scattered around the code that it would be easy to miss something.