Optimize savegame space
The original code saves MAXSTATS amount of static objects, regardless of how many are actually in use, which is quite wasteful, especially when you start adding more members to the struct and/or increase MAXSTATS. So now we only save in use statics.

doorposition[] was originally outside the doorstruct because of the optimized assembly raycaster code, but there's no need to give it separate handling now.

A few objstruct members that were larger than necessary have had their datatypes changed, and finally, the visspot member in statstruct and the transy member in objstruct are a pointless waste of space, so they're gone now.

It may seem silly to save memory now, but not every system comes with GBs of RAM, and mods with exorbitant amounts of objects/actors/etc. can end up bloating save games unnecessarily. It's simple to change, so why not?

Please note that any save games created before these changes will be incompatible now!