Fix trace passing through vertical wall corners
Just a silly typo.

For a short explanation of this bug: xtile/xinttile & ytile/yinttile aren't always the same. xtile by yinttile is a vertical position, and xinttile by ytile is a horizontal position. But if both xtile and ytile incremented by the relevant tilesteps match xinttile and yinttile, it fails to detect any tile there (because the position is before the wall) and just continues tracing until it hits a wall behind it.

In any case, the bug can be easily fixed by forcing xinttile/yinttile to match the xtile/ytile value before testing each entry if the above condition is true. This will ensure that tilemap is being indexed with the correct coordinates.

While the previous workaround used the same condition, this is a lot more straightforward.