Set alpha component in palette ops
This was a serious oversight with the SDL2 implementation; not setting this can cause terrible frame lag when palette shifting, and can cause problems when using SDL_CreateTextureFromSurface as well.

I've heard rumors that Linux builds suffer from frame lag a lot; it'd be much appreciated if somebody could test a build with these changes in Linux to see if they fix it.